requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
exceptiongroup==1.2.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
packaging==24.0
playwright==1.43.0
pluggy==1.5.0
pyee==11.1.0
pytest==8.2.0
pytest-base-url==2.1.0
pytest-playwright==0.5.0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.31.0
sqlparse==0.5.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
